+++ include/mathutils/mathutils.h
@@ -4,8 +4,8 @@
 #include <cmath>
 
 namespace mathutils {
-    float toRadians(float degrees);
-    float toDegrees(float radians);
+    double toRadians(double degrees);
+    double toDegrees(double radians);
 };
 
 #endif // MATHUTIL_H
