+++ src/player.cpp
@@ -1,30 +1,30 @@
 #include "player.h"
 #include "mathutils/mathutils.h"
 
-static const float defaultStartX = 22.0f;
-static const float defaultStartY = 11.5f;
-static const float defaultFov = 66.0f;
-static const float moveSpeed = 3.0f;
-static const float rotationSpeed = 2.0f;
+static const double defaultStartX = 22.0;
+static const double defaultStartY = 11.5;
+static const double defaultFov = 66.0;
+static const double moveSpeed = 3.0;
+static const double rotationSpeed = 2.0;
 
 Player::Player()
     : position(defaultStartX, defaultStartY),
-      direction(-1.0f, 0.0f),
+      direction(-1.0, 0.0),
       velocity(),
-      rotation(0.0f) {
-    float fovRadians = mathutils::toRadians(defaultFov);
-    float focalLength = tan(fovRadians / 2.0f);
+      rotation(0.0) {
+    double fovRadians = mathutils::toRadians(defaultFov);
+    double focalLength = tan(fovRadians / 2.0);
 
     cameraPlane.set(0, focalLength);
 }
 
-Player::Player(float startX, float startY, float fov) 
+Player::Player(double startX, double startY, double fov)
     : position(startX, startY),
-      direction(-1.0f, 0.0f),
+      direction(-1.0, 0.0),
       velocity(),
-      rotation(0.0f) { 
-    float fovRadians = mathutils::toRadians(fov);
-    float focalLength = tan(fovRadians / 2.0f);
+      rotation(0.0) {
+    double fovRadians = mathutils::toRadians(fov);
+    double focalLength = tan(fovRadians / 2.0);
 
     cameraPlane.set(0, focalLength);
 }
@@ -37,16 +37,16 @@ void Player::rotate(Direction direction) {
     rotation = rotationSpeed * direction;
 }
 
-void Player::update(float delta) {
+void Player::update(double delta) {
     direction.rotate(rotation * delta);
     cameraPlane.rotate(rotation * delta);
     position += direction * velocity * delta; 
 }
 
 bool Player::isMovingForwards() const {
-    return velocity.x > 0.0f && velocity.y > 0.0f;
+    return velocity.x > 0.0 && velocity.y > 0.0;
 }
 
 bool Player::isMovingBackwards() const {
-    return velocity.x < 0.0f && velocity.y < 0.0f;
+    return velocity.x < 0.0 && velocity.y < 0.0;
 }
